Feature/spring security

+++ src/main/java/com/haruhiism/bbs/controller/ArticleController.java
@@ -1,31 +1,32 @@
 package com.haruhiism.bbs.controller;
 
-import com.haruhiism.bbs.command.article.*;
-import com.haruhiism.bbs.domain.authentication.LoginSessionInfo;
-import com.haruhiism.bbs.domain.dto.*;
-import com.haruhiism.bbs.exception.auth.AuthenticationFailedException;
+import com.haruhiism.bbs.command.article.ArticleListCommand;
+import com.haruhiism.bbs.domain.dto.BoardCommentDTO;
+import com.haruhiism.bbs.domain.dto.ResourceDTO;
+import com.haruhiism.bbs.domain.entity.BoardAccount;
 import com.haruhiism.bbs.service.article.ArticleService;
 import com.haruhiism.bbs.service.comment.CommentService;
 import com.haruhiism.bbs.service.file.FileHandlerService;
 import com.haruhiism.bbs.service.manage.AccountManagerService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.annotation.CurrentSecurityContext;
+import org.springframework.security.core.context.SecurityContext;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
+import java.security.Principal;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+
+import static com.haruhiism.bbs.domain.dto.BoardArticleDTO.*;
 
 @Controller
 @RequiredArgsConstructor
@@ -39,31 +40,37 @@ public class ArticleController {
     private final FileHandlerService fileHandlerService;
 
 
+    private String getUsernameFromSecurityContext(SecurityContext context) {
+        return isAnonymousUser(context) ? "" : ((BoardAccount) context.getAuthentication().getPrincipal()).getAlias();
+    }
+
+    private boolean isAnonymousUser(SecurityContext context) {
+        return context.getAuthentication() instanceof AnonymousAuthenticationToken;
+    }
+
     @GetMapping("/list")
     public String listBoardArticles(Model model,
                                     @ModelAttribute("command") @Valid ArticleListCommand command,
                                     BindingResult bindingResult,
-                                    HttpServletRequest request){
+                                    Principal principal){ // https://www.baeldung.com/get-user-in-spring-security
 
         if(bindingResult.hasErrors()){
             return "redirect:/board/list";
         }
 
-        BoardArticlesDTO articlesDTO = command.getKeyword().isBlank() ?
+        PagedArticles articles = command.getKeyword().isBlank() ?
                 articleService.readAllByPages(command.getPageNum(), command.getPageSize()) :
                 articleService.searchAllByPages(command.getMode(), command.getKeyword(), command.getPageNum(), command.getPageSize());
 
-        model.addAttribute("articles", articlesDTO.getBoardArticles());
-        model.addAttribute("commentSizes", articlesDTO.getBoardArticleCommentSize());
-        model.addAttribute("totalPages", articlesDTO.getTotalPages());
-        model.addAttribute("currentPage", command.getPageNum());
+        model.addAttribute("articles", articles.getArticles());
+        model.addAttribute("totalPages", articles.getPages());
+        model.addAttribute("currentPage", articles.getCurrentPage());
         model.addAttribute("pageSize", command.getPageSize());
         model.addAttribute("searchMode", command.getMode().name());
         model.addAttribute("searchKeyword", command.getKeyword());
 
-        HttpSession session = request.getSession(false);
         Map<String, String> links = new LinkedHashMap<>();
-        if(session == null){
+        if(principal == null){
             links.put("LOGIN", "/account/login");
             links.put("REGISTER", "/account/register");
         } else {
@@ -71,8 +78,7 @@ public class ArticleController {
             links.put("WITHDRAW", "/account/withdraw");
             links.put("MANAGE", "/account/manage");
 
-            String userId = ((LoginSessionInfo)session.getAttribute("loginSessionInfo")).getUserID();
-            if(accountManagerService.authManagerAccess(userId)){
+            if(accountManagerService.authManagerAccess(principal.getName())){
                 links.put("ADMIN CONSOLE", "/manage/console");
             }
         }
@@ -83,25 +89,22 @@ public class ArticleController {
 
     @GetMapping("/read")
     public String readBoardArticle(Model model,
-                                   @Valid ArticleReadCommand command,
-                                   BindingResult bindingResult,
-                                   HttpServletRequest request) {
+                                   @RequestParam(name = "id") long id,
+                                   @RequestParam(name = "commentPage", defaultValue = "0") int commentPage,
+                                   @CurrentSecurityContext SecurityContext context) {
 
-        if (bindingResult.hasErrors()) {
-            return "redirect:/board/list";
-        }
+        if (id <= 0 || commentPage < 0) return "redirect:/board/list";
 
-        BoardArticleDTO article = articleService.readArticle(command.getId());
-        BoardCommentsDTO comments = commentService.readCommentsOfArticle(article.getId(), command.getCommentPage(), 10);
-        List<ResourceDTO> resources = fileHandlerService.listResourcesOfArticle(command.getId());
+        Read article = articleService.readArticle(id);
+        BoardCommentDTO.PagedComments comments = commentService.readArticleCommentsPaged(id, commentPage, 10);
+        List<ResourceDTO> resources = fileHandlerService.listResourcesOfArticle(id);
 
-        LoginSessionInfo loginSessionInfo = getLoginSessionInfoFromHttpSession(request.getSession(false));
-        model.addAttribute("loginUsername", loginSessionInfo == null ? "" : loginSessionInfo.getUsername());
-
+        model.addAttribute("loginUsername", getUsernameFromSecurityContext(context));
+        model.addAttribute("authorized", article.isAnonymous() || article.getUserId().equals(context.getAuthentication().getName()));
         model.addAttribute("article", article);
-        model.addAttribute("comments", comments.getBoardComments());
+        model.addAttribute("comments", comments.getComments());
         model.addAttribute("currentCommentPage", comments.getCurrentPage());
-        model.addAttribute("totalCommentPages", comments.getTotalPages());
+        model.addAttribute("totalCommentPages", comments.getPages());
         model.addAttribute("resources", resources);
 
         return "board/read";
@@ -111,88 +114,94 @@ public class ArticleController {
     @GetMapping("/write")
     public String writeBoardArticle(
             Model model,
-            @ModelAttribute("command") ArticleSubmitCommand command,
-            HttpServletRequest request){
-
-        LoginSessionInfo loginSessionInfo = getLoginSessionInfoFromHttpSession(request.getSession(false));
-        model.addAttribute("loginUsername", loginSessionInfo == null ? "" : loginSessionInfo.getUsername());
-
+            @ModelAttribute("command") Submit command,
+            @CurrentSecurityContext SecurityContext context){
+        // https://stackoverflow.com/questions/57053736/how-to-check-if-user-is-logged-in-or-anonymous-in-spring-security
+        model.addAttribute("loginUsername", getUsernameFromSecurityContext(context));
         return "board/write";
     }
 
     @PostMapping("/write")
     public String submitBoardArticle(
             Model model,
-            @ModelAttribute("command") @Valid ArticleSubmitCommand command,
+            @ModelAttribute("command") @Validated(Submit.Create.class) Submit command,
             BindingResult bindingResult,
-            HttpServletRequest request){
+            @CurrentSecurityContext SecurityContext context) {
 
-        LoginSessionInfo loginSessionInfo = getLoginSessionInfoFromHttpSession(request.getSession(false));
-
-        if(bindingResult.hasErrors()){
-            model.addAttribute("loginUsername", loginSessionInfo == null ? "" : loginSessionInfo.getUsername());
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("loginUsername", getUsernameFromSecurityContext(context));
             return "/board/write";
         }
 
-        BoardArticleDTO articleDTO = new BoardArticleDTO(command);
-        AuthDTO authDTO = AuthDTO.builder().loginSessionInfo(loginSessionInfo).build();
-        Long createdArticleId = articleService.createArticle(articleDTO, authDTO);
+        Submit articleDTO = Submit.builder()
+                .writer(command.getWriter())
+                .title(command.getTitle())
+                .content(command.getContent())
+                .password(command.getPassword())
+                .build();
+
+        long createdArticleId;
+        if (isAnonymousUser(context)) {
+            createdArticleId = articleService.createArticle(articleDTO);
+        } else {
+            createdArticleId = articleService.createArticle(articleDTO, context.getAuthentication().getName());
+        }
         fileHandlerService.store(command.getUploadedFiles(), createdArticleId);
+
+        return "redirect:/board/read?id=" + createdArticleId;
+    }
+
+    @GetMapping("/edit")
+    public String requestEditArticle(Model model,
+                                     @Validated(Authorize.Request.class) @ModelAttribute("command") Authorize command,
+                                     BindingResult bindingResult,
+                                     Principal principal){
+
+        if(bindingResult.hasErrors()) return "redirect:/board/list";
+
+        Read article = articleService.readArticle(command.getId());
+        if(article.isAnonymous()) return "board/editRequest";
+        if(principal != null && article.getUserId().equals(principal.getName())) {
+            Submit edit = Submit.builder()
+                    .id(article.getId())
+                    .writer(article.getWriter())
+                    .title(article.getTitle())
+                    .content(article.getContent())
+                    .password("PASSWORD_IS_ENCODED").build();
+
+            model.addAttribute("article", edit);
+            model.addAttribute("resources", fileHandlerService.listResourcesOfArticle(command.getId()));
+            return "board/edit";
+        }
 
         return "redirect:/board/list";
     }
 
-
-    private boolean isArticleWrittenByLoggedInAccount(Long articleId){
-        return !articleService.readArticle(articleId).getUserId().isBlank();
-    }
-
-    private void addArticleToModel(Model model, Long articleID, AuthDTO authDTO){
-        model.addAttribute("article", articleService.authArticleEdit(articleID, authDTO).orElseThrow(AuthenticationFailedException::new));
-        model.addAttribute("resources", fileHandlerService.listResourcesOfArticle(articleID));
-    }
-
-
-    @GetMapping("/edit")
-    public String requestEditArticle(Model model,
-                                     @ModelAttribute("command") @Validated(ArticleEditRequestValidationGroup.class) ArticleEditRequestCommand command,
-                                     BindingResult bindingResult,
-                                     HttpServletRequest request){
-
-        if(bindingResult.hasErrors()){
-            return "redirect:/board/list";
-        }
-
-        if (isArticleWrittenByLoggedInAccount(command.getId())) {
-            LoginSessionInfo loginSessionInfo = getLoginSessionInfoFromHttpSession(request.getSession(false));
-            if(loginSessionInfo == null){
-                throw new AuthenticationFailedException();
-            }
-
-            addArticleToModel(model, command.getId(), AuthDTO.builder().loginSessionInfo(loginSessionInfo).build());
-            model.addAttribute("password", "THIS_IS_USELESS_PASSWORD");
-            return "board/edit";
-        } else {
-            return "board/editRequest";
-        }
-    }
-
     @PostMapping("/edit")
     public String authEditArticle(Model model,
-                                  @ModelAttribute("command") @Validated(ArticleEditSubmitValidationGroup.class) ArticleEditRequestCommand command,
+                                  @Validated(Authorize.Submit.class) @ModelAttribute("command") Authorize command,
                                   BindingResult bindingResult){
 
         if (bindingResult.hasErrors()) {
             return "board/editRequest";
         }
 
-        try {
-            addArticleToModel(model, command.getId(), AuthDTO.builder().rawPassword(command.getPassword()).build());
-        } catch (AuthenticationFailedException ex){
+        if(!articleService.authorizeAnonymousArticleAccess(command.getId(), command.getPassword())) {
             bindingResult.addError(new FieldError("command", "password", "Password not matched."));
             return "board/editRequest";
         }
 
+        Read read = articleService.readArticle(command.getId());
+        Submit article = Submit.builder()
+                .id(read.getId())
+                .title(read.getTitle())
+                .writer(read.getWriter())
+                .password(command.getPassword()) // not anonymous: doesn't need password. but anonymous does because of submit auth.
+                .content(read.getContent())
+                .build();
+
+        model.addAttribute("article", article);
+        model.addAttribute("resources", fileHandlerService.listResourcesOfArticle(command.getId()));
         model.addAttribute("password", command.getPassword());
         return "board/edit";
     }
@@ -201,78 +210,56 @@ public class ArticleController {
     @PostMapping("/edit/submit")
     public String submitEditArticle(
             Model model,
-            @ModelAttribute("article") @Valid ArticleEditSubmitCommand command,
+            @ModelAttribute("article") @Validated(Submit.Update.class) Submit command,
             BindingResult bindingResult,
-            HttpServletRequest request){
+            Principal principal){
+
+        Read article = articleService.readArticle(command.getId());
 
         if (bindingResult.hasErrors()) {
-            model.addAttribute("password", command.getPassword());
             model.addAttribute("resources", fileHandlerService.listResourcesOfArticle(command.getId()));
             return "board/edit";
         }
 
-        BoardArticleDTO editedArticleDTO = BoardArticleDTO.builder()
-                .id(command.getId())
-                .title(command.getTitle())
-                .content(command.getContent()).build();
+        if(article.isAnonymous() && articleService.authorizeAnonymousArticleAccess(command.getId(), command.getPassword()) ||
+            principal != null && article.getUserId().equals(principal.getName())) {
+            articleService.updateArticle(command);
+            fileHandlerService.delete(command.getDelete(), command.getId());
+            fileHandlerService.store(command.getUploadedFiles(), command.getId());
+        }
 
-        AuthDTO authDTO = AuthDTO.builder()
-                .rawPassword(command.getPassword())
-                .loginSessionInfo(getLoginSessionInfoFromHttpSession(request.getSession(false))).build();
-
-        articleService.updateArticle(editedArticleDTO, authDTO);
-        fileHandlerService.delete(command.getDelete(), command.getId());
-        fileHandlerService.store(command.getUploadedFiles(), command.getId());
         return "redirect:/board/list";
     }
 
 
     @GetMapping("/remove")
     public String requestRemoveArticle(
-            Model model,
-            @ModelAttribute("command") @Validated(ArticleRemoveRequestValidationGroup.class) ArticleRemoveRequestCommand command,
+            @Validated(Authorize.Request.class) @ModelAttribute("command") Authorize command,
             BindingResult bindingResult,
-            HttpServletRequest request){
+            Principal principal) {
 
-        if (bindingResult.hasErrors()) {
-            return "redirect:/board/list";
-        }
+        if (bindingResult.hasErrors()) return "redirect:/board/list";
 
-        if (isArticleWrittenByLoggedInAccount(command.getId())) {
-            articleService.deleteArticle(
-                    command.getId(),
-                    AuthDTO.builder().loginSessionInfo(getLoginSessionInfoFromHttpSession(request.getSession(false))).build());
-            return "redirect:/board/list";
-        } else {
-            return "board/removeRequest";
-        }
+        Read article = articleService.readArticle(command.getId());
+        if (article.isAnonymous()) return "board/removeRequest";
+        if (principal != null && article.getUserId().equals(principal.getName())) articleService.deleteArticle(command.getId());
+        return "redirect:/board/list";
     }
 
     @PostMapping("/remove")
-    public String authRemoveArticle(@ModelAttribute("command") @Validated(ArticleRemoveSubmitValidationGroup.class) ArticleRemoveRequestCommand command,
+    public String authRemoveArticle(@ModelAttribute("command") @Validated(Authorize.Submit.class) Authorize command,
                                     BindingResult bindingResult){
 
         if(bindingResult.hasErrors()){
             return "board/removeRequest";
         }
 
-        try {
-            articleService.deleteArticle(
-                    command.getId(),
-                    AuthDTO.builder()
-                            .rawPassword(command.getPassword())
-                            .build());
-        } catch (AuthenticationFailedException exception) {
+        if(articleService.authorizeAnonymousArticleAccess(command.getId(), command.getPassword())) {
+            articleService.deleteArticle(command.getId());
+            return "redirect:/board/list";
+        } else {
             bindingResult.addError(new FieldError("command", "password", "Password not matched."));
             return "board/removeRequest";
         }
-
-        return "redirect:/board/list";
-    }
-
-    private LoginSessionInfo getLoginSessionInfoFromHttpSession(HttpSession session) {
-        if(session == null) return null;
-        String sessionAuthAttribute = "loginSessionInfo";
-        return (LoginSessionInfo) session.getAttribute(sessionAuthAttribute);
     }
 }
